psych-chatbot-backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
sentence-transformers

